More small fixes
